First version of warning dialog

+++ src/navigation/menu_bar.py
@@ -354,20 +354,34 @@ class MenuNav:
         ensure_future(coroutine(self))
 
     def do_exit(self) -> None:
-        """Exit"""
-        settings_path = os.path.join(NOTES_DIR, ".user_setting.json")
-        with open(settings_path, "r") as f:
-            user = json.load(f)
+        async def coroutine(self: MenuNav): -> None
+            if self.application_state.current_path:
+                with open(self.current_path) as f:
+                    text = f.read()
+                if text != self.application.text_field:
+                    dialog = ConfirmDialog(title = "Unsaved Changes",
+                                            text = f"The file {self.application_state.current_path} contains unsaved changes. Are you sure you want to exit without saving?"
+                                            )
+                   confirm_exit =  await self.show_dialog_as_float(dialog)
+                else:
+                    confirm_exit = False
+                if confirm_exit:
+                
+                    """Exit"""
+                    settings_path = os.path.join(NOTES_DIR, ".user_setting.json")
+                    with open(settings_path, "r") as f:
+                        user = json.load(f)
 
-        self.application_state.user_settings[
-            "last_path"
-        ] = self.application_state.current_path
+                    self.application_state.user_settings[
+                        "last_path"
+                    ] = self.application_state.current_path
 
-        user["last_path"] = self.application_state.current_path
-        with open(settings_path, "w") as f:
-            json.dump(user, f)
+                    user["last_path"] = self.application_state.current_path
+                    with open(settings_path, "w") as f:
+                        json.dump(user, f)
 
-        get_app().exit()
+                    get_app().exit()
+        ensure_future(coroutine(self))
 
     def do_time_date(self) -> None:
         """Inserts current datetime into self.text_field from menu"""
